Clean up code syntax highlighting
I was using prism.js (https://prismjs.com) to do code highlighting.
This wasn't an intentional decision, the theme I was *ahem*.. influenced
by was using it. When I loaded a page with code, it would first show up
all ugly until the JS ran and formatted it correctly.

Then I started thinking... wait a minute, why am I running JS on the
client to do syntax highlighting? This is a static website!

So now I just use the built in Hugo formatter, which you can read more
about here: https://gohugo.io/content-management/syntax-highlighting/

It's not as fancy but it's static!